ui/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via sanic
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
html5tagger==1.3.0
    # via
    #   sanic
    #   tracerite
httptools==0.6.1
    # via sanic
idna==3.4
    # via requests
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
multidict==6.0.4
    # via sanic
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   sanic-ext
requests==2.31.0
    # via -r requirements.in
sanic[ext]==23.6.0
    # via -r requirements.in
sanic-ext==23.6.0
    # via sanic
sanic-routing==23.6.0
    # via sanic
tracerite==1.1.1
    # via sanic
typing-extensions==4.8.0
    # via sanic
ujson==5.8.0
    # via sanic
urllib3==2.1.0
    # via requests
uvloop==0.19.0
    # via sanic
websockets==12.0
    # via sanic
